#18 :: Fixed misleading use of context module
There was in fact no point in loading roboplot inside context previously.
Indeed, this simply had the effect of loading it and adding it to the list of already imported modules, from where the scripts can get a reference.
Thus, the ``from context import roboplot`` can be replaced with ``import context`` with exactly the same result (thanks to the python module lookup protocol). Further, removing the roboplot import completely in the context module simply has the effect of delaying the module load until the script asks for it (which is a bit clearer).